Rename combatable-character file name
