+++ src/character/index.ts
@@ -1,4 +1,4 @@
 export { default as Character } from './character';
-export { default as CombatableCharacter } from './combatable';
+export { default as CombatableCharacter } from './combatable-character';
 export { default as Hero } from './hero';
 export { default as Monster } from './monster';
